Remove debounce and add to individual implementation as it is not used elsewhere
