+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/price-slider/frontend.ts
@@ -11,7 +11,41 @@ import {
 	ProductFilterPriceContext,
 	ProductFilterPriceStore,
 } from '../price-filter/frontend';
-import { debounce } from '../../utils';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type DebouncedFunction< T extends ( ...args: any[] ) => any > = ( (
+	...args: Parameters< T >
+) => void ) & { flush: () => void };
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const debounce = < T extends ( ...args: any[] ) => any >(
+	func: T,
+	wait: number,
+	immediate?: boolean
+): DebouncedFunction< T > => {
+	let timeout: ReturnType< typeof setTimeout > | null;
+	let latestArgs: Parameters< T > | null = null;
+
+	const debounced = ( ( ...args: Parameters< T > ) => {
+		latestArgs = args;
+		if ( timeout ) clearTimeout( timeout );
+		timeout = setTimeout( () => {
+			timeout = null;
+			if ( ! immediate && latestArgs ) func( ...latestArgs );
+		}, wait );
+		if ( immediate && ! timeout ) func( ...args );
+	} ) as DebouncedFunction< T >;
+
+	debounced.flush = () => {
+		if ( timeout && latestArgs ) {
+			func( ...latestArgs );
+			clearTimeout( timeout );
+			timeout = null;
+		}
+	};
+
+	return debounced;
+};
 
 store( 'woocommerce/product-filter-price-slider', {
 	state: {
