+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/utils/index.ts
@@ -7,4 +7,3 @@ export * from './get-allowed-blocks';
 export * from './get-colors-from-block-supports';
 export * from './get-product-filters-css';
 export * from './price-currency';
-export * from './debounce';
